+++ src/lib.rs
@@ -229,6 +229,8 @@ use ratatui::{
     text::Line,
     widgets::{Block, BorderType, Borders, Widget},
 };
+use widget::inner::TuiLoggerInner;
+use widget::inner::TuiWidgetInnerState;
 
 mod circular;
 #[cfg(feature = "slog-support")]
@@ -247,6 +249,14 @@ pub use crate::slog::TuiSlogDrain;
 pub use crate::tracing_subscriber::TuiTracingSubscriberLayer;
 #[doc(no_inline)]
 pub use log::LevelFilter;
+
+pub mod widget;
+pub use widget::inner::TuiWidgetState;
+pub use widget::smart::TuiLoggerSmartWidget;
+pub use widget::standard::TuiLoggerWidget;
+
+pub mod file;
+pub use file::TuiLoggerFile;
 
 struct ExtLogRecord {
     timestamp: DateTime<Local>,
@@ -359,67 +369,6 @@ struct HotLog {
     mover_thread: Option<thread::JoinHandle<()>>,
 }
 
-/// This closely follows the options of [``TuiLoggerSmartWidget``] but is used of logging to a file.
-pub struct TuiLoggerFile {
-    dump: File,
-    format_separator: char,
-    timestamp_fmt: Option<String>,
-    format_output_target: bool,
-    format_output_file: bool,
-    format_output_line: bool,
-    format_output_level: Option<TuiLoggerLevelOutput>,
-}
-
-impl TuiLoggerFile {
-    pub fn new(fname: &str) -> Self {
-        TuiLoggerFile {
-            dump: OpenOptions::new()
-                .create(true)
-                .append(true)
-                .open(fname)
-                .expect("Failed to open dump File"),
-            format_separator: ':',
-            timestamp_fmt: Some("[%Y:%m:%d %H:%M:%S]".to_string()),
-            format_output_file: true,
-            format_output_line: true,
-            format_output_target: true,
-            format_output_level: Some(TuiLoggerLevelOutput::Long),
-        }
-    }
-    pub fn output_target(mut self, enabled: bool) -> Self {
-        self.format_output_target = enabled;
-        self
-    }
-    pub fn output_file(mut self, enabled: bool) -> Self {
-        self.format_output_file = enabled;
-        self
-    }
-    pub fn output_line(mut self, enabled: bool) -> Self {
-        self.format_output_line = enabled;
-        self
-    }
-    pub fn output_timestamp(mut self, fmt: Option<String>) -> Self {
-        self.timestamp_fmt = fmt;
-        self
-    }
-    pub fn output_separator(mut self, sep: char) -> Self {
-        self.format_separator = sep;
-        self
-    }
-    pub fn output_level(mut self, level: Option<TuiLoggerLevelOutput>) -> Self {
-        self.format_output_level = level;
-        self
-    }
-}
-
-struct TuiLoggerInner {
-    hot_depth: usize,
-    events: CircularBuffer<ExtLogRecord>,
-    dump: Option<TuiLoggerFile>,
-    total_events: usize,
-    default: LevelFilter,
-    targets: LevelConfig,
-}
 struct TuiLogger {
     hot_select: Mutex<HotSelect>,
     hot_log: Mutex<HotLog>,
@@ -710,112 +659,6 @@ pub enum TuiWidgetEvent {
     EscapeKey,
 }
 
-#[derive(Default)]
-struct TuiWidgetInnerState {
-    config: LevelConfig,
-    nr_items: usize,
-    selected: usize,
-    opt_timestamp_bottom: Option<DateTime<Local>>,
-    opt_timestamp_next_page: Option<DateTime<Local>>,
-    opt_timestamp_prev_page: Option<DateTime<Local>>,
-    opt_selected_target: Option<String>,
-    opt_selected_visibility_more: Option<LevelFilter>,
-    opt_selected_visibility_less: Option<LevelFilter>,
-    opt_selected_recording_more: Option<LevelFilter>,
-    opt_selected_recording_less: Option<LevelFilter>,
-    offset: usize,
-    hide_off: bool,
-    hide_target: bool,
-    focus_selected: bool,
-}
-impl TuiWidgetInnerState {
-    pub fn new() -> TuiWidgetInnerState {
-        TuiWidgetInnerState::default()
-    }
-    fn transition(&mut self, event: TuiWidgetEvent) {
-        use TuiWidgetEvent::*;
-        match event {
-            SpaceKey => {
-                self.hide_off ^= true;
-            }
-            HideKey => {
-                self.hide_target ^= true;
-            }
-            FocusKey => {
-                self.focus_selected ^= true;
-            }
-            UpKey => {
-                if !self.hide_target && self.selected > 0 {
-                    self.selected -= 1;
-                }
-            }
-            DownKey => {
-                if !self.hide_target && self.selected + 1 < self.nr_items {
-                    self.selected += 1;
-                }
-            }
-            LeftKey => {
-                if let Some(selected_target) = self.opt_selected_target.take() {
-                    if let Some(selected_visibility_less) = self.opt_selected_visibility_less.take()
-                    {
-                        self.config.set(&selected_target, selected_visibility_less);
-                    }
-                }
-            }
-            RightKey => {
-                if let Some(selected_target) = self.opt_selected_target.take() {
-                    if let Some(selected_visibility_more) = self.opt_selected_visibility_more.take()
-                    {
-                        self.config.set(&selected_target, selected_visibility_more);
-                    }
-                }
-            }
-            PlusKey => {
-                if let Some(selected_target) = self.opt_selected_target.take() {
-                    if let Some(selected_recording_more) = self.opt_selected_recording_more.take() {
-                        set_level_for_target(&selected_target, selected_recording_more);
-                    }
-                }
-            }
-            MinusKey => {
-                if let Some(selected_target) = self.opt_selected_target.take() {
-                    if let Some(selected_recording_less) = self.opt_selected_recording_less.take() {
-                        set_level_for_target(&selected_target, selected_recording_less);
-                    }
-                }
-            }
-            PrevPageKey => self.opt_timestamp_bottom = self.opt_timestamp_prev_page,
-            NextPageKey => self.opt_timestamp_bottom = self.opt_timestamp_next_page,
-            EscapeKey => self.opt_timestamp_bottom = None,
-        }
-    }
-}
-
-/// This struct contains the shared state of a TuiLoggerWidget and a TuiLoggerTargetWidget.
-#[derive(Default)]
-pub struct TuiWidgetState {
-    inner: Arc<Mutex<TuiWidgetInnerState>>,
-}
-impl TuiWidgetState {
-    /// Create a new TuiWidgetState
-    pub fn new() -> TuiWidgetState {
-        TuiWidgetState {
-            inner: Arc::new(Mutex::new(TuiWidgetInnerState::new())),
-        }
-    }
-    pub fn set_default_display_level(self, levelfilter: LevelFilter) -> TuiWidgetState {
-        self.inner.lock().config.default_display_level = Some(levelfilter);
-        self
-    }
-    pub fn set_level_for_target(self, target: &str, levelfilter: LevelFilter) -> TuiWidgetState {
-        self.inner.lock().config.set(target, levelfilter);
-        self
-    }
-    pub fn transition(&mut self, event: TuiWidgetEvent) {
-        self.inner.lock().transition(event);
-    }
-}
-
 /// This is the definition for the TuiLoggerTargetWidget,
 /// which allows configuration of the logger system and selection of log messages.
 pub struct TuiLoggerTargetWidget<'b> {
@@ -1052,648 +895,4 @@ impl<'b> Widget for TuiLoggerTargetWidget<'b> {
 pub enum TuiLoggerLevelOutput {
     Abbreviated,
     Long,
-}
-
-pub struct TuiLoggerWidget<'b> {
-    block: Option<Block<'b>>,
-    /// Base style of the widget
-    style: Style,
-    /// Level based style
-    style_error: Option<Style>,
-    style_warn: Option<Style>,
-    style_debug: Option<Style>,
-    style_trace: Option<Style>,
-    style_info: Option<Style>,
-    format_separator: char,
-    format_timestamp: Option<String>,
-    format_output_level: Option<TuiLoggerLevelOutput>,
-    format_output_target: bool,
-    format_output_file: bool,
-    format_output_line: bool,
-    state: Arc<Mutex<TuiWidgetInnerState>>,
-}
-impl<'b> Default for TuiLoggerWidget<'b> {
-    fn default() -> TuiLoggerWidget<'b> {
-        //TUI_LOGGER.move_events();
-        TuiLoggerWidget {
-            block: None,
-            style: Default::default(),
-            style_error: None,
-            style_warn: None,
-            style_debug: None,
-            style_trace: None,
-            style_info: None,
-            format_separator: ':',
-            format_timestamp: Some("%H:%M:%S".to_string()),
-            format_output_level: Some(TuiLoggerLevelOutput::Long),
-            format_output_target: true,
-            format_output_file: true,
-            format_output_line: true,
-            state: Arc::new(Mutex::new(TuiWidgetInnerState::new())),
-        }
-    }
-}
-impl<'b> TuiLoggerWidget<'b> {
-    pub fn block(mut self, block: Block<'b>) -> Self {
-        self.block = Some(block);
-        self
-    }
-    fn opt_style(mut self, style: Option<Style>) -> Self {
-        if let Some(s) = style {
-            self.style = s;
-        }
-        self
-    }
-    fn opt_style_error(mut self, style: Option<Style>) -> Self {
-        if style.is_some() {
-            self.style_error = style;
-        }
-        self
-    }
-    fn opt_style_warn(mut self, style: Option<Style>) -> Self {
-        if style.is_some() {
-            self.style_warn = style;
-        }
-        self
-    }
-    fn opt_style_info(mut self, style: Option<Style>) -> Self {
-        if style.is_some() {
-            self.style_info = style;
-        }
-        self
-    }
-    fn opt_style_trace(mut self, style: Option<Style>) -> Self {
-        if style.is_some() {
-            self.style_trace = style;
-        }
-        self
-    }
-    fn opt_style_debug(mut self, style: Option<Style>) -> Self {
-        if style.is_some() {
-            self.style_debug = style;
-        }
-        self
-    }
-    pub fn style(mut self, style: Style) -> Self {
-        self.style = style;
-        self
-    }
-    pub fn style_error(mut self, style: Style) -> Self {
-        self.style_error = Some(style);
-        self
-    }
-    pub fn style_warn(mut self, style: Style) -> Self {
-        self.style_warn = Some(style);
-        self
-    }
-    pub fn style_info(mut self, style: Style) -> Self {
-        self.style_info = Some(style);
-        self
-    }
-    pub fn style_trace(mut self, style: Style) -> Self {
-        self.style_trace = Some(style);
-        self
-    }
-    pub fn style_debug(mut self, style: Style) -> Self {
-        self.style_debug = Some(style);
-        self
-    }
-    fn opt_output_separator(mut self, opt_sep: Option<char>) -> Self {
-        if let Some(ch) = opt_sep {
-            self.format_separator = ch;
-        }
-        self
-    }
-    /// Separator character between field.
-    /// Default is ':'
-    pub fn output_separator(mut self, sep: char) -> Self {
-        self.format_separator = sep;
-        self
-    }
-    fn opt_output_timestamp(mut self, opt_fmt: Option<Option<String>>) -> Self {
-        if let Some(fmt) = opt_fmt {
-            self.format_timestamp = fmt;
-        }
-        self
-    }
-    /// The format string can be defined as described in
-    /// <https://docs.rs/chrono/0.4.19/chrono/format/strftime/index.html>
-    ///
-    /// If called with None, timestamp is not included in output.
-    ///
-    /// Default is %H:%M:%S
-    pub fn output_timestamp(mut self, fmt: Option<String>) -> Self {
-        self.format_timestamp = fmt;
-        self
-    }
-    fn opt_output_level(mut self, opt_fmt: Option<Option<TuiLoggerLevelOutput>>) -> Self {
-        if let Some(fmt) = opt_fmt {
-            self.format_output_level = fmt;
-        }
-        self
-    }
-    /// Possible values are
-    /// - TuiLoggerLevelOutput::Long        => DEBUG/TRACE/...
-    /// - TuiLoggerLevelOutput::Abbreviated => D/T/...
-    ///
-    /// If called with None, level is not included in output.
-    ///
-    /// Default is Long
-    pub fn output_level(mut self, level: Option<TuiLoggerLevelOutput>) -> Self {
-        self.format_output_level = level;
-        self
-    }
-    fn opt_output_target(mut self, opt_enabled: Option<bool>) -> Self {
-        if let Some(enabled) = opt_enabled {
-            self.format_output_target = enabled;
-        }
-        self
-    }
-    /// Enables output of target field of event
-    ///
-    /// Default is true
-    pub fn output_target(mut self, enabled: bool) -> Self {
-        self.format_output_target = enabled;
-        self
-    }
-    fn opt_output_file(mut self, opt_enabled: Option<bool>) -> Self {
-        if let Some(enabled) = opt_enabled {
-            self.format_output_file = enabled;
-        }
-        self
-    }
-    /// Enables output of file field of event
-    ///
-    /// Default is true
-    pub fn output_file(mut self, enabled: bool) -> Self {
-        self.format_output_file = enabled;
-        self
-    }
-    fn opt_output_line(mut self, opt_enabled: Option<bool>) -> Self {
-        if let Some(enabled) = opt_enabled {
-            self.format_output_line = enabled;
-        }
-        self
-    }
-    /// Enables output of line field of event
-    ///
-    /// Default is true
-    pub fn output_line(mut self, enabled: bool) -> Self {
-        self.format_output_line = enabled;
-        self
-    }
-    fn inner_state(mut self, state: Arc<Mutex<TuiWidgetInnerState>>) -> Self {
-        self.state = state;
-        self
-    }
-    pub fn state(mut self, state: &TuiWidgetState) -> Self {
-        self.state = state.inner.clone();
-        self
-    }
-    fn format_event(&self, evt: &ExtLogRecord) -> (String, Option<Style>) {
-        let mut output = String::new();
-        let (col_style, lev_long, lev_abbr, with_loc) = match evt.level {
-            log::Level::Error => (self.style_error, "ERROR", "E", true),
-            log::Level::Warn => (self.style_warn, "WARN ", "W", true),
-            log::Level::Info => (self.style_info, "INFO ", "I", true),
-            log::Level::Debug => (self.style_debug, "DEBUG", "D", true),
-            log::Level::Trace => (self.style_trace, "TRACE", "T", true),
-        };
-        if let Some(fmt) = self.format_timestamp.as_ref() {
-            output.push_str(&format!("{}", evt.timestamp.format(fmt)));
-            output.push(self.format_separator);
-        }
-        match &self.format_output_level {
-            None => {}
-            Some(TuiLoggerLevelOutput::Abbreviated) => {
-                output.push_str(lev_abbr);
-                output.push(self.format_separator);
-            }
-            Some(TuiLoggerLevelOutput::Long) => {
-                output.push_str(lev_long);
-                output.push(self.format_separator);
-            }
-        }
-        if self.format_output_target {
-            output.push_str(&evt.target);
-            output.push(self.format_separator);
-        }
-        if with_loc {
-            if self.format_output_file {
-                output.push_str(&evt.file);
-                output.push(self.format_separator);
-            }
-            if self.format_output_line {
-                output.push_str(&format!("{}", evt.line));
-                output.push(self.format_separator);
-            }
-        }
-        (output, col_style)
-    }
-}
-impl<'b> Widget for TuiLoggerWidget<'b> {
-    fn render(mut self, area: Rect, buf: &mut Buffer) {
-        buf.set_style(area, self.style);
-        let list_area = match self.block.take() {
-            Some(b) => {
-                let inner_area = b.inner(area);
-                b.render(area, buf);
-                inner_area
-            }
-            None => area,
-        };
-        let indent = 9;
-        if list_area.width < indent + 4 || list_area.height < 1 {
-            return;
-        }
-
-        let mut state = self.state.lock();
-        let la_height = list_area.height as usize;
-        let mut lines: Vec<(Option<Style>, u16, String)> = vec![];
-        {
-            state.opt_timestamp_next_page = None;
-            let opt_timestamp_bottom = state.opt_timestamp_bottom;
-            let mut opt_timestamp_prev_page = None;
-            let mut tui_lock = TUI_LOGGER.inner.lock();
-            let mut circular = CircularBuffer::new(10); // MAGIC constant
-            for evt in tui_lock.events.rev_iter() {
-                if let Some(level) = state.config.get(&evt.target) {
-                    if level < evt.level {
-                        continue;
-                    }
-                } else if let Some(level) = state.config.default_display_level {
-                    if level < evt.level {
-                        continue;
-                    }
-                }
-                if state.focus_selected {
-                    if let Some(target) = state.opt_selected_target.as_ref() {
-                        if target != &evt.target {
-                            continue;
-                        }
-                    }
-                }
-                // Here all filters have been applied,
-                // So check, if user is paging through history
-                if let Some(timestamp) = opt_timestamp_bottom.as_ref() {
-                    if *timestamp < evt.timestamp {
-                        circular.push(evt.timestamp);
-                        continue;
-                    }
-                }
-                if !circular.is_empty() {
-                    state.opt_timestamp_next_page = circular.take().first().cloned();
-                }
-                let (mut output, col_style) = self.format_event(evt);
-                let mut sublines: Vec<&str> = evt.msg.lines().rev().collect();
-                output.push_str(sublines.pop().unwrap());
-                for subline in sublines {
-                    lines.push((col_style, indent, subline.to_string()));
-                }
-                lines.push((col_style, 0, output));
-                if lines.len() == la_height {
-                    break;
-                }
-                if opt_timestamp_prev_page.is_none() && lines.len() >= la_height / 2 {
-                    opt_timestamp_prev_page = Some(evt.timestamp);
-                }
-            }
-            state.opt_timestamp_prev_page = opt_timestamp_prev_page.or(state.opt_timestamp_bottom);
-        }
-        let la_left = list_area.left();
-        let la_top = list_area.top();
-        let la_width = list_area.width as usize;
-
-        // lines is a vector with bottom line at index 0
-        // wrapped_lines will be a vector with top line first
-        let mut wrapped_lines = CircularBuffer::new(la_height);
-        let rem_width = la_width - indent as usize;
-        while let Some((style, left, line)) = lines.pop() {
-            if line.chars().count() > la_width {
-                wrapped_lines.push((style, left, line.chars().take(la_width).collect()));
-                let mut remain: String = line.chars().skip(la_width).collect();
-                while remain.chars().count() > rem_width {
-                    let remove: String = remain.chars().take(rem_width).collect();
-                    wrapped_lines.push((style, indent, remove));
-                    remain = remain.chars().skip(rem_width).collect();
-                }
-                wrapped_lines.push((style, indent, remain.to_owned()));
-            } else {
-                wrapped_lines.push((style, left, line));
-            }
-        }
-
-        let offset: u16 = if state.opt_timestamp_bottom.is_none() {
-            0
-        } else {
-            let lines_cnt = wrapped_lines.len();
-            (la_height - lines_cnt) as u16
-        };
-
-        for (i, (sty, left, l)) in wrapped_lines.iter().enumerate() {
-            buf.set_stringn(
-                la_left + left,
-                la_top + i as u16 + offset,
-                l,
-                l.len(),
-                sty.unwrap_or(self.style),
-            );
-        }
-    }
-}
-
-/// The Smart Widget combines the TuiLoggerWidget and the TuiLoggerTargetWidget
-/// into a nice combo, where the TuiLoggerTargetWidget can be shown/hidden.
-///
-/// In the title the number of logging messages/s in the whole buffer is shown.
-pub struct TuiLoggerSmartWidget<'a> {
-    title_log: Line<'a>,
-    title_target: Line<'a>,
-    style: Option<Style>,
-    border_style: Style,
-    border_type: BorderType,
-    highlight_style: Option<Style>,
-    style_error: Option<Style>,
-    style_warn: Option<Style>,
-    style_debug: Option<Style>,
-    style_trace: Option<Style>,
-    style_info: Option<Style>,
-    style_show: Option<Style>,
-    style_hide: Option<Style>,
-    style_off: Option<Style>,
-    format_separator: Option<char>,
-    format_timestamp: Option<Option<String>>,
-    format_output_level: Option<Option<TuiLoggerLevelOutput>>,
-    format_output_target: Option<bool>,
-    format_output_file: Option<bool>,
-    format_output_line: Option<bool>,
-    state: Arc<Mutex<TuiWidgetInnerState>>,
-}
-impl<'a> Default for TuiLoggerSmartWidget<'a> {
-    fn default() -> Self {
-        //TUI_LOGGER.move_events();
-        TuiLoggerSmartWidget {
-            title_log: Line::from("Tui Log"),
-            title_target: Line::from("Tui Target Selector"),
-            style: None,
-            border_style: Style::default(),
-            border_type: BorderType::Plain,
-            highlight_style: None,
-            style_error: None,
-            style_warn: None,
-            style_debug: None,
-            style_trace: None,
-            style_info: None,
-            style_show: None,
-            style_hide: None,
-            style_off: None,
-            format_separator: None,
-            format_timestamp: None,
-            format_output_level: None,
-            format_output_target: None,
-            format_output_file: None,
-            format_output_line: None,
-            state: Arc::new(Mutex::new(TuiWidgetInnerState::new())),
-        }
-    }
-}
-impl<'a> TuiLoggerSmartWidget<'a> {
-    pub fn highlight_style(mut self, style: Style) -> Self {
-        self.highlight_style = Some(style);
-        self
-    }
-    pub fn border_style(mut self, style: Style) -> Self {
-        self.border_style = style;
-        self
-    }
-    pub fn border_type(mut self, border_type: BorderType) -> Self {
-        self.border_type = border_type;
-        self
-    }
-    pub fn style(mut self, style: Style) -> Self {
-        self.style = Some(style);
-        self
-    }
-    pub fn style_error(mut self, style: Style) -> Self {
-        self.style_error = Some(style);
-        self
-    }
-    pub fn style_warn(mut self, style: Style) -> Self {
-        self.style_warn = Some(style);
-        self
-    }
-    pub fn style_info(mut self, style: Style) -> Self {
-        self.style_info = Some(style);
-        self
-    }
-    pub fn style_trace(mut self, style: Style) -> Self {
-        self.style_trace = Some(style);
-        self
-    }
-    pub fn style_debug(mut self, style: Style) -> Self {
-        self.style_debug = Some(style);
-        self
-    }
-    pub fn style_off(mut self, style: Style) -> Self {
-        self.style_off = Some(style);
-        self
-    }
-    pub fn style_hide(mut self, style: Style) -> Self {
-        self.style_hide = Some(style);
-        self
-    }
-    pub fn style_show(mut self, style: Style) -> Self {
-        self.style_show = Some(style);
-        self
-    }
-    /// Separator character between field.
-    /// Default is ':'
-    pub fn output_separator(mut self, sep: char) -> Self {
-        self.format_separator = Some(sep);
-        self
-    }
-    /// The format string can be defined as described in
-    /// <https://docs.rs/chrono/0.4.19/chrono/format/strftime/index.html>
-    ///
-    /// If called with None, timestamp is not included in output.
-    ///
-    /// Default is %H:%M:%S
-    pub fn output_timestamp(mut self, fmt: Option<String>) -> Self {
-        self.format_timestamp = Some(fmt);
-        self
-    }
-    /// Possible values are
-    /// - TuiLoggerLevelOutput::Long        => DEBUG/TRACE/...
-    /// - TuiLoggerLevelOutput::Abbreviated => D/T/...
-    ///
-    /// If called with None, level is not included in output.
-    ///
-    /// Default is Long
-    pub fn output_level(mut self, level: Option<TuiLoggerLevelOutput>) -> Self {
-        self.format_output_level = Some(level);
-        self
-    }
-    /// Enables output of target field of event
-    ///
-    /// Default is true
-    pub fn output_target(mut self, enabled: bool) -> Self {
-        self.format_output_target = Some(enabled);
-        self
-    }
-    /// Enables output of file field of event
-    ///
-    /// Default is true
-    pub fn output_file(mut self, enabled: bool) -> Self {
-        self.format_output_file = Some(enabled);
-        self
-    }
-    /// Enables output of line field of event
-    ///
-    /// Default is true
-    pub fn output_line(mut self, enabled: bool) -> Self {
-        self.format_output_line = Some(enabled);
-        self
-    }
-    pub fn title_target<T>(mut self, title: T) -> Self
-    where
-        T: Into<Line<'a>>,
-    {
-        self.title_target = title.into();
-        self
-    }
-    pub fn title_log<T>(mut self, title: T) -> Self
-    where
-        T: Into<Line<'a>>,
-    {
-        self.title_log = title.into();
-        self
-    }
-    pub fn state(mut self, state: &TuiWidgetState) -> Self {
-        self.state = state.inner.clone();
-        self
-    }
-}
-impl<'a> Widget for TuiLoggerSmartWidget<'a> {
-    /// Nothing to draw for combo widget
-    fn render(self, area: Rect, buf: &mut Buffer) {
-        let entries_s = {
-            let mut tui_lock = TUI_LOGGER.inner.lock();
-            let first_timestamp = tui_lock
-                .events
-                .iter()
-                .next()
-                .map(|entry| entry.timestamp.timestamp_millis());
-            let last_timestamp = tui_lock
-                .events
-                .rev_iter()
-                .next()
-                .map(|entry| entry.timestamp.timestamp_millis());
-            if let Some(first) = first_timestamp {
-                if let Some(last) = last_timestamp {
-                    let dt = last - first;
-                    if dt > 0 {
-                        tui_lock.events.len() as f64 / (dt as f64) * 1000.0
-                    } else {
-                        0.0
-                    }
-                } else {
-                    0.0
-                }
-            } else {
-                0.0
-            }
-        };
-
-        let mut title_log = self.title_log.clone();
-        title_log
-            .spans
-            .push(format!(" [log={:.1}/s]", entries_s).into());
-
-        let hide_target = self.state.lock().hide_target;
-        if hide_target {
-            let tui_lw = TuiLoggerWidget::default()
-                .block(
-                    Block::default()
-                        .title(title_log)
-                        .border_style(self.border_style)
-                        .border_type(self.border_type)
-                        .borders(Borders::ALL),
-                )
-                .opt_style(self.style)
-                .opt_style_error(self.style_error)
-                .opt_style_warn(self.style_warn)
-                .opt_style_info(self.style_info)
-                .opt_style_debug(self.style_debug)
-                .opt_style_trace(self.style_trace)
-                .opt_output_separator(self.format_separator)
-                .opt_output_timestamp(self.format_timestamp)
-                .opt_output_level(self.format_output_level)
-                .opt_output_target(self.format_output_target)
-                .opt_output_file(self.format_output_file)
-                .opt_output_line(self.format_output_line)
-                .inner_state(self.state);
-            tui_lw.render(area, buf);
-        } else {
-            let mut width: usize = 0;
-            {
-                let hot_targets = &TUI_LOGGER.inner.lock().targets;
-                let mut state = self.state.lock();
-                let hide_off = state.hide_off;
-                {
-                    let targets = &mut state.config;
-                    targets.merge(hot_targets);
-                    for (t, levelfilter) in targets.iter() {
-                        if hide_off && levelfilter == &LevelFilter::Off {
-                            continue;
-                        }
-                        width = width.max(t.len())
-                    }
-                }
-            }
-            let chunks = Layout::default()
-                .direction(Direction::Horizontal)
-                .constraints(vec![
-                    Constraint::Length(width as u16 + 6 + 2),
-                    Constraint::Min(10),
-                ])
-                .split(area);
-            let tui_ltw = TuiLoggerTargetWidget::default()
-                .block(
-                    Block::default()
-                        .title(self.title_target)
-                        .border_style(self.border_style)
-                        .border_type(self.border_type)
-                        .borders(Borders::ALL),
-                )
-                .opt_style(self.style)
-                .opt_highlight_style(self.highlight_style)
-                .opt_style_off(self.style_off)
-                .opt_style_hide(self.style_hide)
-                .opt_style_show(self.style_show)
-                .inner_state(self.state.clone());
-            tui_ltw.render(chunks[0], buf);
-            let tui_lw = TuiLoggerWidget::default()
-                .block(
-                    Block::default()
-                        .title(title_log)
-                        .border_style(self.border_style)
-                        .border_type(self.border_type)
-                        .borders(Borders::ALL),
-                )
-                .opt_style(self.style)
-                .opt_style_error(self.style_error)
-                .opt_style_warn(self.style_warn)
-                .opt_style_info(self.style_info)
-                .opt_style_debug(self.style_debug)
-                .opt_style_trace(self.style_trace)
-                .opt_output_separator(self.format_separator)
-                .opt_output_timestamp(self.format_timestamp)
-                .opt_output_level(self.format_output_level)
-                .opt_output_target(self.format_output_target)
-                .opt_output_file(self.format_output_file)
-                .opt_output_line(self.format_output_line)
-                .inner_state(self.state.clone());
-            tui_lw.render(chunks[1], buf);
-        }
-    }
 }
