+++ src/widget/mod.rs
@@ -1,0 +1,3 @@
+pub mod inner;
+pub mod smart;
+pub mod standard;
